# apps/server/assets/files/swagger.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "Conduit API",
		"description": "Conduit API documentation",
		"contact": {
			"name": "RealWorld",
			"url": "https://realworld.how"
		},
		"license": {
			"name": "MIT License",
			"url": "https://opensource.org/licenses/MIT"
		},
		"version": "1.0.0"
	},
...
	"servers": [
		{
			"url": "/api"
		}
	],
	"paths": {
...
		"/articles": {
			"get": {
				"tags": ["Articles"],
				"summary": "Get recent articles globally",
				"description": "Get most recent articles globally. Use query parameters to filter results. Auth is optional",
				"operationId": "GetArticles",
				"parameters": [
					{
						"name": "tag",
						"in": "query",
						"description": "Filter by tag",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "author",
						"in": "query",
						"description": "Filter by author (username)",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "favorited",
						"in": "query",
						"description": "Filter by favorites of a user (username)",
						"schema": {
							"type": "string"
						}
					},
					{
						"$ref": "#/components/parameters/offsetParam"
					},
					{
						"$ref": "#/components/parameters/limitParam"
					}
				],
				"responses": {
					"200": {
						"$ref": "#/components/responses/MultipleArticlesResponse"
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"422": {
						"$ref": "#/components/responses/GenericError"
					}
				}
			},
			"post": {
				"tags": ["Articles"],
				"summary": "Create an article",
				"description": "Create an article. Auth is required",
				"operationId": "CreateArticle",
				"requestBody": {
					"$ref": "#/components/requestBodies/NewArticleRequest"
				},
				"responses": {
					"201": {
						"$ref": "#/components/responses/SingleArticleResponse"
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"422": {
						"$ref": "#/components/responses/GenericError"
					}
				},
				"security": [
					{
						"Token": []
					}
				],
				"x-codegen-request-body-name": "article"
			}
		},
		"/articles/{slug}": {
			"get": {
				"tags": ["Articles"],
				"summary": "Get an article",
				"description": "Get an article. Auth not required",
				"operationId": "GetArticle",
				"parameters": [
					{
						"name": "slug",
						"in": "path",
						"description": "Slug of the article to get",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"$ref": "#/components/responses/SingleArticleResponse"
					},
					"422": {
						"$ref": "#/components/responses/GenericError"
					}
				}
			},
			"put": {
				"tags": ["Articles"],
				"summary": "Update an article",
				"description": "Update an article. Auth is required",
				"operationId": "UpdateArticle",
				"parameters": [
					{
						"name": "slug",
						"in": "path",
						"description": "Slug of the article to update",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"$ref": "#/components/requestBodies/UpdateArticleRequest"
				},
				"responses": {
					"200": {
						"$ref": "#/components/responses/SingleArticleResponse"
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"422": {
						"$ref": "#/components/responses/GenericError"
					}
				},
				"security": [
					{
						"Token": []
					}
				],
				"x-codegen-request-body-name": "article"
			},
			"delete": {
				"tags": ["Articles"],
				"summary": "Delete an article",
				"description": "Delete an article. Auth is required",
				"operationId": "DeleteArticle",
				"parameters": [
					{
						"name": "slug",
						"in": "path",
						"description": "Slug of the article to delete",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"$ref": "#/components/responses/EmptyOkResponse"
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"422": {
						"$ref": "#/components/responses/GenericError"
					}
				},
				"security": [
					{
						"Token": []
					}
				]
			}
		},
...
		"/tags": {
			"get": {
				"tags": ["Tags"],
				"summary": "Get tags",
				"description": "Get tags. Auth not required",
				"operationId": "GetTags",
				"responses": {
					"200": {
						"$ref": "#/components/responses/TagsResponse"
					},
					"422": {
						"$ref": "#/components/responses/GenericError"
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
...
			"Profile": {
				"required": ["bio", "following", "image", "username"],
				"type": "object",
				"properties": {
					"username": {
						"type": "string"
					},
					"bio": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"following": {
						"type": "boolean"
					}
				}
			},
			"Article": {
				"required": [
					"author",
					"body",
					"createdAt",
					"description",
					"favorited",
					"favoritesCount",
					"slug",
					"tagList",
					"title",
					"updatedAt"
				],
				"type": "object",
				"properties": {
					"slug": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"body": {
						"type": "string"
					},
					"tagList": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"favorited": {
						"type": "boolean"
					},
					"favoritesCount": {
						"type": "integer"
					},
					"author": {
						"$ref": "#/components/schemas/Profile"
					}
				}
			},
			"NewArticle": {
				"required": ["body", "description", "title"],
				"type": "object",
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"body": {
						"type": "string"
					},
					"tagList": {
						"type": "array",
						"items": {
							"type": "string"
						}
					}
				}
			},
			"UpdateArticle": {
				"type": "object",
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"body": {
						"type": "string"
					}
				}
			},
...
			"GenericErrorModel": {
				"required": ["errors"],
				"type": "object",
				"properties": {
					"errors": {
						"required": ["body"],
						"type": "object",
						"properties": {
							"body": {
								"type": "array",
								"items": {
									"type": "string"
								}
							}
						}
					}
				}
			}
		},
		"responses": {
			"TagsResponse": {
				"description": "Tags",
				"content": {
					"application/json": {
						"schema": {
							"required": ["tags"],
							"type": "object",
							"properties": {
								"tags": {
									"type": "array",
									"items": {
										"type": "string"
									}
								}
							}
						}
					}
				}
			},
...
			"SingleArticleResponse": {
				"description": "Single article",
				"content": {
					"application/json": {
						"schema": {
							"required": ["article"],
							"type": "object",
							"properties": {
								"article": {
									"$ref": "#/components/schemas/Article"
								}
							}
						}
					}
				}
			},
			"MultipleArticlesResponse": {
				"description": "Multiple articles",
				"content": {
					"application/json": {
						"schema": {
							"required": ["articles", "articlesCount"],
							"type": "object",
							"properties": {
								"articles": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Article"
									}
								},
								"articlesCount": {
									"type": "integer"
								}
							}
						}
					}
				}
			},
...
			"EmptyOkResponse": {
				"description": "No content",
				"content": {}
			},
			"Unauthorized": {
				"description": "Unauthorized",
				"content": {}
			},
			"GenericError": {
				"description": "Unexpected error",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/GenericErrorModel"
						}
					}
				}
			}
		},
		"requestBodies": {
...
			"NewArticleRequest": {
				"required": true,
				"description": "Article to create",
				"content": {
					"application/json": {
						"schema": {
							"required": ["article"],
							"type": "object",
							"properties": {
								"article": {
									"$ref": "#/components/schemas/NewArticle"
								}
							}
						}
					}
				}
			},
			"UpdateArticleRequest": {
				"required": true,
				"description": "Article to update",
				"content": {
					"application/json": {
						"schema": {
							"required": ["article"],
							"type": "object",
							"properties": {
								"article": {
									"$ref": "#/components/schemas/UpdateArticle"
								}
							}
						}
					}
				}
			},
...
		},
		"parameters": {
			"offsetParam": {
				"in": "query",
				"name": "offset",
				"required": false,
				"schema": {
					"type": "integer",
					"minimum": 0
				},
				"description": "The number of items to skip before starting to collect the result set."
			},
			"limitParam": {
				"in": "query",
				"name": "limit",
				"required": false,
				"schema": {
					"type": "integer",
					"minimum": 1,
					"default": 20
				},
				"description": "The numbers of items to return."
			}
		},
		"securitySchemes": {
			"Token": {
				"type": "apiKey",
				"description": "For accessing the protected API resources, you must have received a a valid JWT token after registering or logging in. This JWT token must then be used for all protected resources by passing it in via the 'Authorization' header.\n\nA JWT token is generated by the API by either registering via /users or logging in via /users/login.\n\nThe following format must be in the 'Authorization' header :\n\n    Token xxxxxx.yyyyyyy.zzzzzz\n    \n",
				"name": "Authorization",
				"in": "header"
			}
		}
	}
}